Move tests to their own module, use flatten_text and better regexes.
