+++ pydoctor/test/test_astbuilder.py
@@ -1,4 +1,4 @@
-import re
+
 from typing import Optional, Tuple, Type, overload, cast
 import ast
 import textwrap
@@ -2021,62 +2021,3 @@ def test_variable_named_like_current_module(systemcls: Type[model.System]) -> No
     example = True
     ''', systemcls=systemcls, modname="example")
     assert 'example' in mod.contents
-
-def test_twisted_python_deprecate(capsys: CapSys) -> None:
-    """
-    It recognizes Twisted deprecation decorators and add the
-    deprecation info as part of the documentation.
-    """
-    from . import test_templatewriter
-
-    # Adjusted from Twisted's tests at
-    # https://github.com/twisted/twisted/blob/3bbe558df65181ed455b0c5cc609c0131d68d265/src/twisted/python/test/test_release.py#L516
-
-    docstring = "text in docstring"
-    privateDocstring = "should also appear in output"
-
-    mod = fromText("from twisted.python.deprecate import deprecated\n"
-            "from incremental import Version\n"
-            "@deprecated(Version('Twisted', 15, 0, 0), "
-            "'Baz')\n"
-            "def foo():\n"
-            "    '{}'\n"
-            "from twisted.python import deprecate\n"
-            "import incremental\n"
-            "@deprecate.deprecated(incremental.Version('Twisted', 16, 0, 0))\n"
-            "def _bar():\n"
-            "    '{}'\n"
-            "@deprecated(Version('Twisted', 14, 2, 3), replacement='stuff')\n"
-            "class Baz:\n"
-            "    pass\n"
-            "class stuff: ..."
-            "".format(docstring, privateDocstring), systemcls=deprecate.System, modname='mod')
-
-    mod_html = test_templatewriter.getHTMLOf(mod)
-    class_html = test_templatewriter.getHTMLOf(mod.contents['Baz'])
-
-    # assert not capsys.readouterr().out
-    assert not capsys.readouterr().err
-
-    _html_template_with_replacement = r'(.+){name}(.+) was deprecated in {package} {version}; please use (.+){replacement}(.+) instead\.(.+)'
-    _html_template_without_replacement = r'(.+){name}(.+) was deprecated in {package} {version}\.(.+)'
-    
-    assert docstring in mod_html
-    assert privateDocstring in mod_html
-
-    assert re.match(_html_template_with_replacement.format(
-        name='foo', package='Twisted', version=r'15\.0\.0', replacement='Baz'
-    ), mod_html, re.DOTALL), mod_html
-    assert re.match(_html_template_without_replacement.format(
-        name='_bar', package='Twisted', version=r'16\.0\.0'
-    ), mod_html, re.DOTALL), mod_html
-
-    _class = mod.contents['Baz']
-    assert len(_class.extra_info)==1
-    assert re.match(_html_template_with_replacement.format(
-        name='Baz', package='Twisted', version=r'14\.2\.3', replacement='stuff'
-    ), flatten(_class.extra_info[0].to_stan(mod.docstring_linker, False)).strip(), re.DOTALL)
-
-    assert re.match(_html_template_with_replacement.format(
-        name='Baz', package='Twisted', version=r'14\.2\.3', replacement='stuff'
-    ), class_html, re.DOTALL), class_html
